+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# v1.1.0
+> Cleanup
+
+- Removed the `CountdownCtrl` in favor of an extending the link function.
+- Added support for setting when the warning style class is applied via `warnSeconds` attribute.
+- Added support for setting when the danger style class is applied via `dangerSeconds` attribute.
+
+
 # v1.0.1
 > Refactor